Separates calls to catalog service
